Impliment enh #3: 'Add layer to canvas' chkbox added

+++ src/converter_dialog.py
@@ -19,18 +19,17 @@
  ***************************************************************************/
 """
 try:
-    from osgeo import ogr, osr
+    from osgeo import ogr, osr,  gdal
 except ImportError:
-    import ogr, osr
+    import ogr, osr,  gdal
 
 import sys
-import datetime
 from os import path
-from osm_geocoder import OsmGeocoder
-from google_geocoder import GoogleGeocoder
 
 from PyQt4.QtGui import QDialog, QMessageBox, QFileDialog
 from PyQt4.QtCore import QObject, SIGNAL, QString
+
+from qgis.core import QgsVectorLayer,  QgsMapLayerRegistry
 
 from ui_converter_dialog import Ui_ConverterDialog
 
@@ -42,22 +41,19 @@ class ConverterDialog(QDialog, Ui_ConverterDialog):
         #SIGNALS
         QObject.connect(self.btnSelectCsv, SIGNAL("clicked()"), self.select_csv)
         QObject.connect(self.btnSelectShp, SIGNAL("clicked()"), self.select_shp)
-        QObject.connect(self, SIGNAL("accepted()"), self.processing)
+        QObject.connect(self.buttonBox, SIGNAL("accepted()"), self.processing)
 
 
     def select_shp(self):
         shp_path = QString()
         if self.txtShpPath.text():
             shp_path = unicode(self.txtShpPath.text())
-
-
         file_name = QFileDialog.getSaveFileName(self,
                                                 self.tr('Select output SHP file'),
                                                 shp_path,
                                                 self.tr('SHP files (*.shp *.shp)'))
         if not file_name.isEmpty():
             self.txtShpPath.setText(file_name)
-
 
 
     def select_csv(self):
@@ -71,32 +67,46 @@ class ConverterDialog(QDialog, Ui_ConverterDialog):
 
 
     def processing(self):
+        #check user input
+        in_path = unicode(self.txtCsvPath.text()).encode("utf-8")        
         out_path = unicode(self.txtShpPath.text()).encode("utf-8")
-        drv = ogr.GetDriverByName("ESRI Shapefile")
-        #check exists
-        if path.exists(out_path):
-            #message
-            drv.DeleteDataSource(out_path)
-        #create output shp file
-        try:
-            drv = ogr.GetDriverByName("ESRI Shapefile")
-            output_data_source = drv.CreateDataSource(out_path)
-        except:
-            QMessageBox.critical(self, self.tr("RuGeocoder error"),
-                                     self.tr("Output SHP file can't be created!\r%1: %2")
-                                     .arg(unicode(sys.exc_info()[0].__name__))
-                                     .arg(unicode(sys.exc_info()[1])))
+        
+        if not in_path:
+            self.__show_err(self.tr("Select input CSV file!"))
             return
- 
+        if not out_path:
+            self.__show_err(self.tr("Select name for output SHP file!"))
+            return
+        if not path.exists(in_path):
+            self.__show_err(self.tr("Selected CSV file not found!"))
+            return
+        
+        #prepare output data source
+        drv = ogr.GetDriverByName("ESRI Shapefile")
+        #check output datasource exists
+        if path.exists(out_path):
+            if QMessageBox.question(self,  self.tr("RuGeocoder"),  
+                                    self.tr("SHP file with the same name already exists.\r Do you want to overwrite it?"), 
+                                    QMessageBox.Yes | QMessageBox.Cancel) == QMessageBox.Yes:
+                drv.DeleteDataSource(out_path)
+            else:
+                return
+        
+        #create output shp file
+        gdal.ErrorReset()
+        output_data_source = drv.CreateDataSource(out_path)
+        if output_data_source==None:
+            self.__show_err(self.tr("Output SHP file can't be created!\r%1")
+                                     .arg(unicode(gdal.GetLastErrorMsg())))
+            return
 
         wgs_sr = osr.SpatialReference()
         wgs_sr.ImportFromEPSG(4326)
 
-        layer_name = path.basename(out_path).strip(".shp")
+        layer_name = path.splitext(path.basename(out_path))[0]
         output_layer = output_data_source.CreateLayer( layer_name, srs = wgs_sr, geom_type = ogr.wkbPoint )
 
         #copy fields
-        in_path = unicode(self.txtCsvPath.text()).encode("utf-8")
         input_data_source = ogr.Open(in_path)
         csv_layer = input_data_source[0]
 
@@ -106,34 +116,22 @@ class ConverterDialog(QDialog, Ui_ConverterDialog):
             if field_def.GetType()==ogr.OFTString:
                 field_def.SetWidth(300)
             if output_layer.CreateField(field_def) != 0:
-                print "Can't create field %s" % field_def.GetNameRef()
+                self.__show_err( self.tr("Unable to create a field %1!").arg(field_def.GetNameRef()))
                 return
 
         #add geocoder additional fields
         out_defs = output_layer.GetLayerDefn()
         if out_defs.GetFieldIndex("settlement") < 0:
-            field_def = ogr.FieldDefn("settlement", ogr.OFTString)
-            field_def.SetWidth( 300 )
-            if output_layer.CreateField ( field_def ) != 0:
-                print "Can't create field %s" % field_def.GetNameRef()
+            if not self.__add_field(output_layer, "settlement", ogr.OFTString,  300):
                 return
         if out_defs.GetFieldIndex("street") < 0:
-            field_def = ogr.FieldDefn( "street", ogr.OFTString )
-            field_def.SetWidth( 300 )
-            if output_layer.CreateField (field_def) != 0:
-                print "Can't create field %s" % field_def.GetNameRef()
+            if not self.__add_field(output_layer, "street", ogr.OFTString,  300):
                 return
         if out_defs.GetFieldIndex("house_num") < 0:
-            field_def = ogr.FieldDefn("house_num", ogr.OFTString)
-            field_def.SetWidth( 300 )
-            if output_layer.CreateField(field_def) != 0:
-                print "Can't create field %s" % field_def.GetNameRef()
+            if not self.__add_field(output_layer, "house_num", ogr.OFTString,  300):
                 return
         if out_defs.GetFieldIndex("geocoded") < 0:
-            field_def = ogr.FieldDefn("geocoded", ogr.OFTString)
-            field_def.SetWidth(300)
-            if output_layer.CreateField (field_def) != 0:
-                print "Can't create field %s" % field_def.GetNameRef()
+            if not self.__add_field(output_layer, "geocoded", ogr.OFTString,  300):
                 return
 
         in_feat = csv_layer.GetNextFeature()
@@ -142,17 +140,43 @@ class ConverterDialog(QDialog, Ui_ConverterDialog):
             #copy fields
             res = out_feat.SetFrom(in_feat)
             if res != 0:
-                print "Can't copy feature"
-                continue
+                self.__show_err(self.tr("Unable to construct the feature!"))
+                return
             #set geom
             pt = ogr.Geometry(ogr.wkbPoint)
             pt.SetPoint_2D(0, 0, 0)
             out_feat.SetGeometry(pt)
-
+            #add to layer
             if output_layer.CreateFeature(out_feat) != 0:
-                print "Failed to create feature in shapefile.\n"
-                sys.exit(1)
-
+                self.__show_err(self.tr("Failed to create feature in SHP file!"))
+                return
             in_feat = csv_layer.GetNextFeature()
 
         QMessageBox.information(self, self.tr("RuGeocoder"), self.tr("Converting successfully completed"))
+        if self.chkAddToCanvas.isChecked():
+            self.add_layer_to_canvas(out_path)
+        self.close()
+        
+    def add_layer_to_canvas(self, shp_path):
+        if not path.exists(shp_path):
+            return False#message???
+        layer_name = path.splitext(path.basename(shp_path))[0]
+        vector_layer = QgsVectorLayer(shp_path, layer_name, "ogr")
+        if vector_layer.isValid():
+            QgsMapLayerRegistry.instance().addMapLayer(vector_layer)
+            return True
+        else:
+            return False
+        
+    def __add_field(self,  layer, field_name,   field_type=ogr.OFTString,  field_len=None):
+        field_def = ogr.FieldDefn(field_name, field_type)
+        if field_len:
+            field_def.SetWidth(field_len)
+        if layer.CreateField (field_def) != 0:
+            self.__show_err( self.tr("Unable to create a field %1!").arg(field_def.GetNameRef()))
+            return False
+        else:
+            return True
+            
+    def __show_err(self,  msg):
+         QMessageBox.critical(self, self.tr("RuGeocoder error"), msg)
